refactor(config): hoist TS config evaluator into zfb-config-loader leaf crate
`ServerBuilder::config_path()` with a `.ts` path hard-errored at `build()`
because the TS evaluator lived in the `zfb` bin crate, which `zfb-server`
cannot call (`zfb` depends on `zfb-server`, not the reverse).

Extract the evaluator into a new leaf crate `zfb-config-loader` that depends
on neither `zfb` nor `zfb-server`. Both now depend on it:

- New crate owns: `load_from_ts_file` (esbuild bundle + in-process V8 /
  node subprocess), the `embed_v8` feature gate, `output_bounded*` bounded
  subprocess helpers, esbuild resolver wrapper, plugin path/bare-specifier
  resolution (`node_resolve`), and the `config-loader.mjs` / stub JS files.
  It is config-shape-agnostic: returns the evaluated `default` export as a
  `serde_json::Value` + resolved plugin specifiers (`LoadedTsConfig`), so it
  needs neither crate's `Config` struct.
- `zfb::config` wraps it: injects the `EMBEDDED_VENDOR`-backed esbuild getter,
  deserialises the value into the strongly-typed `Config`, applies resolved
  plugins, validates. Public `load_from_dir` API and `LoadOptions` unchanged.
- `zfb-server::embed::load_embed_config` now evaluates `.ts` (on a dedicated
  current-thread runtime, so the sync `build()` works inside or outside a
  tokio runtime) instead of returning the milestone hard-error.

The `embed_v8` feature propagates `zfb` / `zfb-server` -> `zfb-config-loader`
-> `zfb-render`, keeping V8 inclusion decided in one place per consumer.

Closes #1037.
