+++ crates/zfb-config-loader/js/zfb-config-stub.mjs
