+++ crates/zfb-config-loader/tests/fixtures/node_resolve/exports-object/src/pkg-exports-obj/x.js
