+++ crates/zfb-config-loader/tests/fixtures/node_resolve/exports-string/src/pkg-exports-str/x.js
