+++ crates/zfb-config-loader/tests/fixtures/node_resolve/main-field/src/pkg-main/package.json
