+++ crates/zfb-config-loader/tests/fixtures/node_resolve/parent-walk/src/pkg-parent/package.json
