+++ crates/zfb-config-loader/tests/fixtures/node_resolve/real-plugin-shape/src/@takazudo/zfb-plugin-example/package.json
