+++ crates/zfb-config-loader/tests/fixtures/node_resolve/scoped-package/src/@scope/pkg/index.js
